feat: Add environment variable to deployment and job cleanup function
Updated the Kubernetes deployment script and YAML file. An environment variable was added to the deployment configuration. Also, a new function for cleaning up jobs was implemented in the Kubernetes executor. This function waits for a job to finish before deleting it. Additionally, updated documentation with instructions on how to retrieve the latest logs from an application running in Kubernetes.