feat: Add cleanup cron job and update role bindings
A new CronJob has been added to clean up successful jobs every 30 minutes. The service account for the CronJob is specified, along with the command it runs.

The RoleBinding file has been updated to include a new Role and RoleBinding for job cleanup. This allows the service account to get, list, and delete jobs in the default namespace. Additionally, an existing RoleBinding was updated to reference this new role.