feat(executor): add an input parameter
Enhanced the executor script to accept an additional input parameter. This is used to provide data for code execution in different languages (Python, Lua, Rust). The input is also added to the ExecutionPayload struct and passed correctly in k8s_executor.rs. Logging of successful code execution has been added in server.rs.